+++ apps/spu-ui/src/app/_components/plans/schemas/common.ts
@@ -17,3 +17,23 @@ export const proposalSchema = z
     regexPatterns.proposal,
     "Proposal must be exactly 8 digits long and contain only numbers",
   );
+
+export const acquireTimeSchema = z.coerce.number().positive();
+
+/**
+ * sampleTag is a common field used in acquisition related forms and
+ * can be used to generate filenames for acquired data.
+ */
+export const sampleTagSchema = z
+  .string()
+  .min(1, "Sample tag must not be empty")
+  .max(100, "Sample tag must not exceed 100 characters")
+  .regex(
+    regexPatterns.noEmptySpaces,
+    "Sample tag must not contain empty spaces",
+  )
+  .regex(regexPatterns.noDots, "Sample tag must not contain dots")
+  .regex(
+    regexPatterns.invalidChars,
+    'Sample tag must not contain any of the following characters: \\ / : * ? " < > |',
+  );