+++ apps/spu-ui/src/app/_components/plans/setup1-complete-acquisition-form.tsx
@@ -37,12 +37,16 @@ import {
   trayRows,
 } from "~/app/_components/store/setup1/constants";
 import { useSampleStore } from "../store/setup1/use-sample-store";
-import { proposalSchema } from "./schemas/common";
+import {
+  acquireTimeSchema,
+  proposalSchema,
+  sampleTagSchema,
+} from "./schemas/common";
 
 export const planName = "setup1_complete_standard_acquisition";
 
 export const planSchema = z.object({
-  acquireTime: z.coerce.number().positive(),
+  acquireTime: acquireTimeSchema,
   numExposures: z.coerce.number().int().positive(),
   row: z
     .string()
@@ -70,9 +74,7 @@ export const planSchema = z.object({
     ),
   volume: z.coerce.number().positive(),
   proposal: proposalSchema,
-  sampleTag: z
-    .string()
-    .min(1, "Sample name or other form of identification is required"),
+  sampleTag: sampleTagSchema,
   sampleType: z
     .string()
     .transform((val) => val.trimStart().trimEnd())
@@ -85,7 +87,6 @@ export const planSchema = z.object({
   measureUvNumber: z.coerce.number().int().min(0).optional(),
   temperature: z.coerce.number().positive().optional(),
   setTemperature: z.boolean().optional(),
-  bufferTag: z.string().optional(),
   standardOption: z.enum(cleaningOptions).optional(),
   agentsList: z.array(z.string()).optional(),
   agentsDuration: z.array(z.coerce.number().positive()).optional(),
@@ -125,7 +126,6 @@ export function CompleteAcquisitionForm({
         col: sampleParams.col,
         sampleType: sampleParams.sampleType,
         sampleTag: sampleParams.sampleTag,
-        bufferTag: sampleParams.bufferTag,
       }),
     },
   });
@@ -361,32 +361,6 @@ export function CompleteAcquisitionForm({
                 <InfoTooltip>
                   <FieldDescription>
                     The name or other form of identification for the sample.
-                  </FieldDescription>
-                </InfoTooltip>
-              </FieldLabel>
-              <InputGroup>
-                <InputGroupInput {...field} />
-                {fieldState.invalid && (
-                  <InputGroupAddon align={"inline-end"}>
-                    <InfoTooltip variant={"destructive"}>
-                      {fieldState.error?.message}
-                    </InfoTooltip>
-                  </InputGroupAddon>
-                )}
-              </InputGroup>
-            </Field>
-          )}
-        />
-        <Controller
-          control={form.control}
-          name="bufferTag"
-          render={({ field, fieldState }) => (
-            <Field>
-              <FieldLabel>
-                Buffer Tag
-                <InfoTooltip>
-                  <FieldDescription>
-                    The name or other form of identification for the buffer.
                   </FieldDescription>
                 </InfoTooltip>
               </FieldLabel>
@@ -646,7 +620,72 @@ export function CompleteAcquisitionForm({
         />
       </FieldGroup>
       <Separator />
-      <FieldGroup
+      <Controller
+        name="standardOption"
+        control={form.control}
+        render={({ field, fieldState }) => (
+          <Field data-invalid={fieldState.invalid}>
+            <FieldLabel htmlFor={field.name}>
+              Cleaning Option
+              <InfoTooltip>
+                <FieldDescription>
+                  The cleaning preset after acquisition.
+                </FieldDescription>
+              </InfoTooltip>
+            </FieldLabel>
+            <Select onValueChange={field.onChange} defaultValue={field.value}>
+              <SelectTrigger className="w-full" size="sm">
+                <SelectValue placeholder="Select cleaning option" />
+                {fieldState.invalid && (
+                  <InfoTooltip variant={"destructive"}>
+                    {fieldState.error?.message}
+                  </InfoTooltip>
+                )}
+              </SelectTrigger>
+              <SelectContent>
+                {cleaningOptions
+                  .filter((option) => option !== "custom")
+                  .map((option) => {
+                    return (
+                      <SelectItem key={option} value={option}>
+                        {option}
+                      </SelectItem>
+                    );
+                  })}
+              </SelectContent>
+            </Select>
+          </Field>
+        )}
+      />
+      <Separator />
+      <Controller
+        control={form.control}
+        name="proposal"
+        render={({ field, fieldState }) => (
+          <Field>
+            <FieldLabel>
+              Proposal
+              <InfoTooltip>
+                <FieldDescription>
+                  The proposal associated with the sample acquisition.
+                </FieldDescription>
+              </InfoTooltip>
+            </FieldLabel>
+            <InputGroup>
+              <InputGroupInput {...field} />
+              {fieldState.invalid && (
+                <InputGroupAddon align={"inline-end"}>
+                  <InfoTooltip variant={"destructive"}>
+                    {fieldState.error?.message}
+                  </InfoTooltip>
+                </InputGroupAddon>
+              )}
+            </InputGroup>
+          </Field>
+        )}
+      />
+
+      {/* <FieldGroup
         id="cleaning-parameters"
         className={cn("grid w-full grid-flow-row grid-cols-3 gap-2", className)}
       >
@@ -716,7 +755,7 @@ export function CompleteAcquisitionForm({
             </Field>
           )}
         />
-      </FieldGroup>
+      </FieldGroup> */}
       <Button
         type="submit"
         disabled={form.formState.isSubmitting}