+++ apps/spu-ui/src/app/_components/store/setup1/constants.ts
@@ -14,17 +14,6 @@ export const trayColumns = [
 ] as const;
 export const trayRows = ["A", "B", "C", "D", "E", "F", "G", "H"] as const;
 export const trayOptions = ["Tray1", "Tray2"] as const;
-export const acquireTimeOptions = [
-  "200",
-  "100",
-  "50",
-  "25",
-  "12.5",
-  "6.25",
-  "3.125",
-  "1.5625",
-  "0.5",
-] as const;
 export const sampleTypeOptions = ["buffer", "sample"] as const;
 export const cleaningOptions = ["light", "normal", "heavy", "custom"] as const;
 export const standardCleaningOptions = ["light", "normal", "heavy"] as const;