+++ apps/spu-ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sophys-web/spu-ui
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [d3d6296]
+- Updated dependencies [583d6b4]
+- Updated dependencies [fdade43]
+  - @sophys-web/api@0.4.0
+  - @sophys-web/widgets@0.6.0
+  - @sophys-web/api-client@0.4.0
+
+## 0.5.0
+
+### Minor Changes
+
+- 0d859eb: Remove capillary acquisition step from the upload csv form.
+
 ## 0.4.4
 
 ### Patch Changes